+++ examples/example_blog/src/app.rs
@@ -6,7 +6,7 @@ use std::{
 use leptos::prelude::*;
 use leptos_fetch::{QueryClient, QueryDevtools, QueryOptions, QueryScope};
 
-use crate::blog_list::BlogList;
+use crate::{blog_list::BlogList, testcases::TestCases};
 
 pub fn shell(options: LeptosOptions) -> impl IntoView {
     view! {
@@ -80,5 +80,6 @@ pub fn RefetchExample(refetch_enabled: RwSignal<bool>) -> impl IntoView {
                 </p>
             </Suspense>
         </div>
+        <TestCases />
     }
 }